% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allen_cady_procedure.R
\name{allen_cady}
\alias{allen_cady}
\title{Allen-Cady Linear Model Feature Selection}
\usage{
allen_cady(.data, .outcome_var, p_threshold = 0.2)
}
\arguments{
\item{.data}{A data frame, data frame extension (e.g. a tibble), or a lazy data frame (e.g. from dbplyr or dtplyr).}

\item{.outcome_var}{The outcome (dependent) variable for the linear model; must be a column in the data frame.}

\item{p_threshold}{The threshold significance value. Predictors with p-values above the threshold are iteratively removed from the model.}
}
\value{
A list of \emph{i} linear model outputs (e.g. a list of the results from \code{base::lm()}) where \emph{i} is the number of iterations to create a model where all features are below the p-value threshold.
}
\description{

}
\examples{
allen_cady(mtcars, mpg)

}
