% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cdr_plus_nacc_ftld.R
\name{calculate_cdr_plus_nacc_ftld}
\alias{calculate_cdr_plus_nacc_ftld}
\title{Calculate Global CDR plus NACC FTLD Rating}
\usage{
calculate_cdr_plus_nacc_ftld(
  data,
  cdr_variables = c("MEMORY", "ORIENT", "JUDGMENT", "COMMUN", "HOMEHOBB", "PERSCARE",
    "COMPORT", "CDRLANG")
)
}
\arguments{
\item{data}{dataframe object, preferably with the standard NACC variable names}

\item{cdr_variables}{vector of 8 variables used to calculate the CDR plus NACC FTLD score; default argument uses standard NACC variable names}
}
\value{
dataframe
}
\description{
Calculate Global CDR plus NACC FTLD Rating
}
\examples{
calculate_cdr_plus_nacc_ftld(nacc_data)
}
