% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_current_past_difference.R
\name{calculate_current_past_difference}
\alias{calculate_current_past_difference}
\title{Calculate difference between vectors of columns
This function accepts any number of 2 sets of equal length columns, and calculates the difference (that is, past-current = difference).
Intended to be used in conjunction with the combine_current_and_past_observations() function.}
\usage{
calculate_current_past_difference(dataframe, final, initial)
}
\arguments{
\item{dataframe}{dataframe object}

\item{final}{vector of final value columns from dataframe}

\item{initial}{vector of initial value columns from dataframe}
}
\value{
dataframe
}
\description{
Calculate difference between vectors of columns
This function accepts any number of 2 sets of equal length columns, and calculates the difference (that is, past-current = difference).
Intended to be used in conjunction with the combine_current_and_past_observations() function.
}
