% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_mac_and_uds_encounters.R
\name{combine_mac_uds_encounters}
\alias{combine_mac_uds_encounters}
\title{Combine MAC and UDS Encounters}
\usage{
combine_mac_uds_encounters(
  dataframe,
  questions_to_combine,
  day_range = 30,
  DCDate = DCDate,
  PIDN = PIDN
)
}
\arguments{
\item{dataframe}{dataframe object}

\item{questions_to_combine}{select the variables that should be combined. This will be at least the item-level variables, and potentially your summary scores}

\item{day_range}{range of days over which MAC and UDS visits can be considered the "same" visit; default of 30 days}

\item{DCDate}{date identifier; defaults to DCDate variable}

\item{PIDN}{person identifier; defaults to PIDN}
}
\value{
dataframe
}
\description{
This function allows you to combine MAC and UDS encounters for a particular test where the
encounters have been separated into different rows, but really occurred around the same time period.
You can determine what range of days qualifies as a "single visit" using the day_range argument.
}
\section{Examples}{

}

