% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_fuzzy_join.R
\name{custom_fuzzy_join}
\alias{custom_fuzzy_join}
\title{Title}
\usage{
custom_fuzzy_join(
  df1,
  DCDate1,
  PIDN1 = PIDN,
  df2,
  DCDate2,
  PIDN2 = PIDN,
  mode = "left",
  interval = 90
)
}
\arguments{
\item{df1}{first dataframe argument}

\item{DCDate1}{first date argument}

\item{PIDN1}{first ID argument; defaults to PIDN}

\item{df2}{second dataframe argument}

\item{DCDate2}{second date argument}

\item{PIDN2}{second ID argument; defaults to PIDN}

\item{mode}{used to specify if the join should be "left", "right", "full", "semi", or "anti"; defaults to "left"}

\item{interval}{absolute value of days over which to match visits; default of +/-90 days}
}
\value{

}
\description{
Title
}
