% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_for_first_visit.R
\name{filter_for_first_visit}
\alias{filter_for_first_visit}
\title{Title}
\usage{
filter_for_first_visit(.data, id, date)
}
\arguments{
\item{.data}{A data frame, data frame extension (e.g. a tibble), or a lazy data frame (e.g. from dbplyr or dtplyr).}

\item{id}{grouping ID variable}

\item{date}{date variable}
}
\value{

}
\description{
Title
}
\section{Examples}{

}

