% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_for_most_recent_visit.R
\name{filter_for_most_recent_visit}
\alias{filter_for_most_recent_visit}
\title{Title}
\usage{
filter_for_most_recent_visit(dataframe, id, date)
}
\arguments{
\item{date}{}
}
\value{

}
\description{
Title
}
\examples{


}
