% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_if_all_na.R
\name{filter_if_all_na}
\alias{filter_if_all_na}
\title{Remove Rows with Multiple NAs}
\usage{
filter_if_all_na(.data, ...)
}
\arguments{
\item{.data}{A data frame, data frame extension (e.g. a tibble), or a lazy data frame (e.g. from dbplyr or dtplyr).}

\item{...}{One or more unquoted expressions separated by commas. Variable names can be used as if they were positions in the data frame, so expressions like x:y can be used to select a range of variables.}
}
\value{

}
\description{
This function filters out rows where all variables specified in the \code{...} argument are NA.
}
\section{Examples}{

}

