% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_if_any_na.R
\name{filter_if_any_na}
\alias{filter_if_any_na}
\title{Title}
\usage{
filter_if_any_na(dataframe, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
This function filters out rows where any variables specified in the \code{...} argument are NA.
}
\examples{

}
