% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_missing_score.R
\name{impute_missing_score}
\alias{impute_missing_score}
\title{Impute for Missing Values}
\usage{
impute_missing_score(
  .data,
  vars_to_impute,
  scale_name,
  missing_threshold = 0.2
)
}
\arguments{
\item{.data}{A data frame, data frame extension (e.g. a tibble), or a lazy data frame (e.g. from dbplyr or dtplyr).}

\item{vars_to_impute}{vector of variables from which a score will be calculated/imputed}

\item{scale_name}{output name of the scale being calculated}

\item{missing_threshold}{maximum proportion of the scale questions allowed to be missing (i.e. the scale will not be calculated if the number of missing questions is greater than this value)}

\item{toggle_warning}{}
}
\value{

}
\description{
This function calculates imputed scores where item-level data is missing, and allows for the threshold of proportion of missing questions to be adjusted
}
\section{Examples}{

}

