% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Impute Missing Values.R
\name{impute_missing_values}
\alias{impute_missing_values}
\title{Impute for Missing Values
This function calculates imputed scores where item-level data is missing, and allows for the threshold of proportion of missing questions to be adjusted}
\usage{
impute_missing_values(
  dataframe,
  vars_to_impute,
  scale_name,
  missing_threshold = 0.2,
  toggle_warning = TRUE
)
}
\arguments{
\item{dataframe}{dataframe object}

\item{vars_to_impute}{vector of variables from which a score will be calculated/imputed}

\item{scale_name}{output name of the scale being calculated}

\item{missing_threshold}{maximum proportion of the scale questions allowed to be missing (i.e. the scale will not be calculated if the number of missing questions is greater than this value)}

\item{toggle_warning}{}
}
\value{

}
\description{
Impute for Missing Values
This function calculates imputed scores where item-level data is missing, and allows for the threshold of proportion of missing questions to be adjusted
}
