% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_by_id_date_difference.R
\name{join_key_date_difference}
\alias{join_key_date_difference}
\title{Fuzzy Join by Days}
\usage{
join_key_date_difference(
  data1,
  data2,
  id1,
  id2,
  date1,
  date2,
  mode = "left",
  interval = 90
)
}
\arguments{
\item{data1}{first data frame}

\item{data2}{second data frame}

\item{id1}{primary key variable from data1}

\item{id2}{primary key variable from data2}

\item{date1}{date key variable from data1. This column should be of class "Date," but accepts any numeric value}

\item{date2}{date key variable from data2. This column should be of class "Date," but accepts any numeric value}

\item{mode}{specify join type should be "left", "right", "full", "semi", or "anti"; defaults to "left"}

\item{interval}{absolute value of days over which to match observations by id; default of +/-90 days}
}
\value{
data frame comprised of data1 and data2
}
\description{

}
