% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nacc_cdr_data_simulated.R
\docType{data}
\name{nacc_cdr_data_simulated}
\alias{nacc_cdr_data_simulated}
\title{NACC CDR Data Simulated}
\format{
A tibble with 10 columns, which are:
\describe{
\item{NACCID}{description of varaible}
\item{VISITDATE}{description of varaible}
\item{MEMORY}{description of varaible}
\item{ORIENT}{description of varaible}
\item{JUDGMENT}{description of varaible}
\item{COMMUN}{description of varaible}
\item{HOMEHOBB}{description of varaible}
\item{PERSCARE}{description of varaible}
\item{COMPORT}{description of varaible}
\item{CDRLANG}{description of varaible}
}
}
\usage{
nacc_cdr_data_simulated
}
\description{
A data frame simulating the 8 NACC CDR variables that are required to calculate the CDR® plus NACC FTLD score
}
\keyword{datasets}
