% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_negative_values_to_na.R
\name{negative_values_to_na}
\alias{negative_values_to_na}
\title{Negative Values to NA}
\usage{
negative_values_to_na(.data, ...)
}
\arguments{
\item{.data}{A data frame, data frame extension (e.g. a tibble), or a lazy data frame (e.g. from dbplyr or dtplyr).}

\item{...}{any numeric variables where negative values should be set to NA}
}
\value{

}
\description{
Allows for the user to define any numeric variables they would like to change to NA if the value of that variable is less than 0
}
\section{Examples}{

}

