% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtrics_dart_item_score.R
\name{qualtrics_dart_item_score}
\alias{qualtrics_dart_item_score}
\title{Title}
\usage{
qualtrics_dart_item_score(.data, var, value)
}
\arguments{
\item{.data}{A data frame, data frame extension (e.g. a tibble), or a lazy data frame (e.g. from dbplyr or dtplyr).}

\item{value}{}
}
\value{

}
\description{
Title
}
\section{Examples}{

}

