% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtrics_hot_spot_score.R
\name{qualtrics_hot_spot_score}
\alias{qualtrics_hot_spot_score}
\title{Title}
\usage{
qualtrics_hot_spot_score(.data, var)
}
\arguments{
\item{.data}{A data frame, data frame extension (e.g. a tibble), or a lazy data frame (e.g. from dbplyr or dtplyr).}

\item{var}{}
}
\value{

}
\description{
Title
}
\section{Examples}{

}

