% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtrics_sdr_item_response.R
\name{qualtrics_sdr_item_response}
\alias{qualtrics_sdr_item_response}
\title{Title}
\usage{
qualtrics_sdr_item_response(df, variable)
}
\arguments{
\item{variable}{}
}
\value{

}
\description{
Title
}
