% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtrics_sdr_item_score.R
\name{qualtrics_sdr_item_score}
\alias{qualtrics_sdr_item_score}
\title{Title}
\usage{
qualtrics_sdr_item_score(data, two_point1, two_point2 = NA, one_point)
}
\arguments{
\item{one_point}{}
}
\value{

}
\description{
Title
}
