% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtrics_tom_item_response.R
\name{qualtrics_tom_item_response}
\alias{qualtrics_tom_item_response}
\title{Title}
\usage{
qualtrics_tom_item_response(df, root_variable)
}
\arguments{
\item{root_variable}{}
}
\value{

}
\description{
Title
}
