% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_csv_by_string.R
\name{read_csv_by_string_search}
\alias{read_csv_by_string_search}
\title{Read CSV By String Search
This convenience function searches either the default directory or a specified directory for a (1) .csv file that at least partially matches the string_search variable. The file is then read-in to R.}
\usage{
read_csv_by_string_search(string_search, file_parent)
}
\arguments{
\item{string_search}{string to search for in the directory}

\item{file_parent}{can specify the parent directory of the file if needed}
}
\value{

}
\description{
Read CSV By String Search
This convenience function searches either the default directory or a specified directory for a (1) .csv file that at least partially matches the string_search variable. The file is then read-in to R.
}
