% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Remove Rows with Multiple NAs.R
\name{remove_rows_with_mult_NAs}
\alias{remove_rows_with_mult_NAs}
\title{Remove Rows with Multipls NAs
This function allows you to remove rows if ALL specified columns are NA}
\usage{
remove_rows_with_mult_NAs(dataframe, columns)
}
\arguments{
\item{dataframe}{dataframe object}

\item{columns}{variables to check for missingness}
}
\value{

}
\description{
Remove Rows with Multipls NAs
This function allows you to remove rows if ALL specified columns are NA
}
