% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_scoring.R
\name{reverse_scoring}
\alias{reverse_scoring}
\title{Reverse Scoring
Allows for reverse-scoring of any specified columns}
\usage{
reverse_scoring(dataframe, reversed_columns, extremum_max)
}
\arguments{
\item{dataframe}{dataframe object}

\item{reversed_columns}{variables to have their score reversed}

\item{extremum_max}{maximum attainable score on the variables selected; needs to be the same maximum for all vars}
}
\value{

}
\description{
Reverse Scoring
Allows for reverse-scoring of any specified columns
}
