% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_tests_by_name.R
\name{select_by_test_name}
\alias{select_by_test_name}
\title{Title}
\usage{
select_by_test_name(dataframe, test_names, ...)
}
\arguments{
\item{dataframe}{dataframe object}

\item{test_names}{character vector of all names of tests that one would like to select from the dataframe; assumes that the tests share a common string (e.g. "_DART" or "_BISBAS")}

\item{...}{any additional columns to be selected from the dataframe}
}
\value{

}
\description{
Title
}
