% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_variable_naming.R
\name{add_test_name_to_vars}
\alias{add_test_name_to_vars}
\title{Add Test Name to Variables
Adds a name to all variables in the dataframe; can specify if the name appears at the beginning or end of the variables}
\usage{
add_test_name_to_vars(dataframe, test_name, location = 0)
}
\arguments{
\item{dataframe}{dataframe object}

\item{test_name}{name to be added to all variables; must be quoted (e.g. "name")}

\item{location}{where to append the test_name; 0 = end of variables, 1 = beginning of variables; defaults to 0}
}
\value{

}
\description{
Add Test Name to Variables
Adds a name to all variables in the dataframe; can specify if the name appears at the beginning or end of the variables
}
