% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_current_past_difference.R
\name{calculate_current_past_difference}
\alias{calculate_current_past_difference}
\title{Calculate difference between vectors of columns
This function accepts any number of 2 sets of equal length columns, and calculates the difference (that is, past-current = difference).
Intended to be used in conjunction with the combine_current_and_past_observations() function.}
\usage{
calculate_current_past_difference(
  dataframe,
  variables_past,
  variables_current,
  variables_output_prefix,
  PIDN = PIDN,
  DCDate = DCDate
)
}
\arguments{
\item{dataframe}{dataframe object}

\item{variables_past}{variables where the observation is in the past}

\item{variables_current}{variables where the observation is current}

\item{variables_output_prefix}{add a meaningful prefix to the name of the difference columns}

\item{PIDN}{defaults to PIDN; can be any other ID}

\item{DCDate}{defaults to DCDate; can be any other date variable}
}
\value{
dataframe
}
\description{
Calculate difference between vectors of columns
This function accepts any number of 2 sets of equal length columns, and calculates the difference (that is, past-current = difference).
Intended to be used in conjunction with the combine_current_and_past_observations() function.
}
