% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_current_and_past_observations.R
\name{combine_current_and_past_observations}
\alias{combine_current_and_past_observations}
\title{Combine Current and Past Observations}
\usage{
combine_current_and_past_observations(
  dataframe,
  CurrentPast = CurrentPast,
  DCDate = DCDate,
  PIDN = PIDN
)
}
\arguments{
\item{dataframe}{dataframe object}

\item{CurrentPast}{variable indicating whether the observation is a "Current" or "Past" record; default column name expected is CurrentPast, but can be replaced with any column with "current" and "past"/"before" characters}

\item{DCDate}{date column; defaults to DCDate}

\item{PIDN}{identifier column; defaults to PIDN}
}
\value{
dataframe
}
\description{

}
