% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_if_all_na.R
\name{filter_if_all_na}
\alias{filter_if_all_na}
\title{Remove Rows with Multiple NAs}
\usage{
filter_if_all_na(dataframe, ...)
}
\arguments{
\item{dataframe}{dataframe object}
}
\value{

}
\description{
This function filters out rows where all variables specified in the \code{...} argument are NA.
}
