% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_negative_values_to_na.R
\name{negative_values_to_na}
\alias{negative_values_to_na}
\title{Negative Values to NA
Allows for the user to define any numeric variables they would like to change to NA if the value of that variable is less than 0}
\usage{
negative_values_to_na(dataframe, vars_to_clean)
}
\arguments{
\item{dataframe}{dataframe object}

\item{vars_to_clean}{any numeric variables where negative values should be set to NA}
}
\value{

}
\description{
Negative Values to NA
Allows for the user to define any numeric variables they would like to change to NA if the value of that variable is less than 0
}
