% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtrics_dart_item_score.R
\name{qualtrics_dart_item_score}
\alias{qualtrics_dart_item_score}
\title{Title}
\usage{
qualtrics_dart_item_score(data, var, value)
}
\arguments{
\item{value}{}
}
\value{

}
\description{
Title
}
