% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtrics_tom_item_response.R
\name{qualtrics_tom_item_response}
\alias{qualtrics_tom_item_response}
\title{Title}
\usage{
qualtrics_tom_item_response(.data, root_variable)
}
\arguments{
\item{.data}{A data frame, data frame extension (e.g. a tibble), or a lazy data frame (e.g. from dbplyr or dtplyr).}

\item{root_variable}{}
}
\value{

}
\description{
Title
}
\section{Examples}{

}

