% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtrics_ucsf_api_credentials.R
\name{qualtrics_ucsf_api_credentials}
\alias{qualtrics_ucsf_api_credentials}
\title{Title}
\usage{
qualtrics_ucsf_api_credentials(
  api_key,
  base_url = "ucsf.co1.qualtrics.com",
  overwrite = FALSE,
  install = FALSE
)
}
\arguments{
\item{install}{}
}
\value{

}
\description{
Title
}
