% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_csv_by_string.R
\name{read_csv_by_string_search}
\alias{read_csv_by_string_search}
\title{Read CSV By String Search}
\usage{
read_csv_by_string_search(string_search, directory = here::here())
}
\arguments{
\item{string_search}{string or regular expression to search for in the directory}

\item{directory}{string of the directory in which to search for CSV file; defaults to the current working directory using here::here()}
}
\value{

}
\description{
This convenience function searches either the default directory or a specified directory for a (1) .csv file that at least partially matches the string_search variable. The file is then read-in to R.
}
\section{Examples}{

}

