% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_scoring.R
\name{reverse_scoring}
\alias{reverse_scoring}
\title{Reverse Scoring}
\usage{
reverse_scoring(.data, ..., max_score)
}
\arguments{
\item{.data}{A data frame, data frame extension (e.g. a tibble), or a lazy data frame (e.g. from dbplyr or dtplyr).}

\item{...}{variables to have their score reversed}

\item{max_score}{maximum attainable score on the variables selected; needs to be the same maximum for all vars}
}
\value{

}
\description{
This function wraps \code{dplyr::mutate()} to reverse score columns given in the \code{...} argument. This procedure is frequently used for e.g. Likert scale questions whereby different questions within the same questionnaire have positive or negative valence.
}
