% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_tests_by_name.R
\name{select_by_test_name}
\alias{select_by_test_name}
\title{Title}
\usage{
select_by_test_name(.data, test_names, ...)
}
\arguments{
\item{.data}{A data frame, data frame extension (e.g. a tibble), or a lazy data frame (e.g. from dbplyr or dtplyr).}

\item{test_names}{character vector of all names of tests that one would like to select from .data; assumes that the tests share a common string (e.g. "_DART" or "_BISBAS")}

\item{...}{any additional columns to be selected from .data}
}
\value{

}
\description{

}
\section{Examples}{

}

